+++ siumai-core/src/types/chat/request.rs
@@ -298,7 +298,7 @@ impl ChatRequestBuilder {
     }
 
     /// Set the temperature (0.0 to 2.0)
-    pub fn temperature(mut self, temperature: f32) -> Self {
+    pub fn temperature(mut self, temperature: f64) -> Self {
         self.common_params.temperature = Some(temperature);
         self
     }
@@ -310,7 +310,7 @@ impl ChatRequestBuilder {
     }
 
     /// Set the top_p sampling parameter
-    pub fn top_p(mut self, top_p: f32) -> Self {
+    pub fn top_p(mut self, top_p: f64) -> Self {
         self.common_params.top_p = Some(top_p);
         self
     }
