+++ siumai-provider-anthropic/src/providers/anthropic/builder.rs
@@ -48,7 +48,7 @@ impl AnthropicBuilder {
     }
 
     /// Common parameter setting methods (similar to `OpenAI`)
-    pub const fn temperature(mut self, temp: f32) -> Self {
+    pub const fn temperature(mut self, temp: f64) -> Self {
         self.common_params.temperature = Some(temp);
         self
     }
@@ -58,7 +58,7 @@ impl AnthropicBuilder {
         self
     }
 
-    pub const fn top_p(mut self, top_p: f32) -> Self {
+    pub const fn top_p(mut self, top_p: f64) -> Self {
         self.common_params.top_p = Some(top_p);
         self
     }
