+++ siumai-provider-groq/src/providers/groq/builder.rs
@@ -48,7 +48,7 @@ impl GroqBuilder {
         self
     }
 
-    pub fn temperature(mut self, temperature: f32) -> Self {
+    pub fn temperature(mut self, temperature: f64) -> Self {
         self.inner = self.inner.temperature(temperature);
         self
     }
@@ -58,7 +58,7 @@ impl GroqBuilder {
         self
     }
 
-    pub fn top_p(mut self, top_p: f32) -> Self {
+    pub fn top_p(mut self, top_p: f64) -> Self {
         self.inner = self.inner.top_p(top_p);
         self
     }
