+++ siumai-provider-ollama/src/providers/ollama/builder.rs
@@ -54,7 +54,7 @@ impl OllamaBuilder {
     ///
     /// # Arguments
     /// * `temperature` - Temperature value (0.0 to 2.0)
-    pub const fn temperature(mut self, temperature: f32) -> Self {
+    pub const fn temperature(mut self, temperature: f64) -> Self {
         self.common_params.temperature = Some(temperature);
         self
     }
@@ -72,7 +72,7 @@ impl OllamaBuilder {
     ///
     /// # Arguments
     /// * `top_p` - Top-p value (0.0 to 1.0)
-    pub const fn top_p(mut self, top_p: f32) -> Self {
+    pub const fn top_p(mut self, top_p: f64) -> Self {
         self.common_params.top_p = Some(top_p);
         self
     }
