+++ siumai-provider-openai/src/providers/openai/config.rs
@@ -124,7 +124,7 @@ impl OpenAiConfig {
     ///
     /// # Arguments
     /// * `temperature` - The temperature value
-    pub const fn with_temperature(mut self, temperature: f32) -> Self {
+    pub const fn with_temperature(mut self, temperature: f64) -> Self {
         self.common_params.temperature = Some(temperature);
         self
     }
