+++ siumai-registry/src/provider/siumai_builder.rs
@@ -150,7 +150,7 @@ impl SiumaiBuilder {
     }
 
     /// Set temperature
-    pub const fn temperature(mut self, temperature: f32) -> Self {
+    pub const fn temperature(mut self, temperature: f64) -> Self {
         self.common_params.temperature = Some(temperature);
         self
     }
@@ -162,7 +162,7 @@ impl SiumaiBuilder {
     }
 
     /// Set top_p (nucleus sampling parameter)
-    pub const fn top_p(mut self, top_p: f32) -> Self {
+    pub const fn top_p(mut self, top_p: f64) -> Self {
         self.common_params.top_p = Some(top_p);
         self
     }
